requirements.txt
onnxruntime
sentence-transformers
pandas
python-docx
PyPDF2
python-dotenv
scikit-learn
loguru
requests
google-generativeai
aiohttp>=3.8.0
asyncio
bs4
pyyaml
streamlit
typesense>=0.17.0
meilisearch>=0.26.0
opensearch-py>=2.0.0
pyarrow>=12.0.0 